restructure and reorganize components
