+++ packages/web/components/Setup/SetupDone.tsx
@@ -1,8 +1,7 @@
 import { MetaButton, MetaHeading } from '@metafam/ds';
+import { FlexContainer } from 'components/Container';
 import { useRouter } from 'next/router';
 import React from 'react';
-
-import { FlexContainer } from './Container';
 
 export const SetupDone: React.FC = () => {
   const router = useRouter();
