+++ packages/web/components/Setup/SetupPersonality.tsx
@@ -1,8 +1,7 @@
 import { MetaButton, MetaHeading } from '@metafam/ds';
+import { FlexContainer } from 'components/Container';
+import { SetupContext } from 'contexts/SetupContext';
 import React, { useContext } from 'react';
-
-import { SetupContext } from '../contexts/SetupContext';
-import { FlexContainer } from './Container';
 
 export const SetupPersonality: React.FC = () => {
   const { useProgress } = useContext(SetupContext);
