+++ packages/web/pages/_app.tsx
@@ -1,9 +1,8 @@
 import { ChakraProvider, CSSReset, MetaTheme } from '@metafam/ds';
 import { PageHeader } from 'components/PageHeader';
+import { Web3ContextProvider } from 'contexts/Web3Context';
 import { AppProps } from 'next/app';
 import Head from 'next/head';
-
-import { Web3ContextProvider } from '../contexts/Web3Context';
 
 const app: React.FC<AppProps> = ({ pageProps, Component }) => {
   return (
@@ -11,6 +10,7 @@ const app: React.FC<AppProps> = ({ pageProps, Component }) => {
       <CSSReset />
       <Head>
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>MetaGame</title>
       </Head>
       <Web3ContextProvider>
         {!pageProps.hidePageHeader && <PageHeader />}
